feat: Add rule to prevent staging secret files
This commit introduces a new rule, `NoSecretFilesStaged`, to prevent the staging of files that might contain secrets. The rule checks staged and modified files against a list of regular expressions to identify potential secret files. If any are found, it raises an error.

The commit also adds the new rule to `GitAdd` and `GitCommit` events.

Also the commit fixes a typo.